view 생성

+++ src/main/java/service/BoardServiceImpl.java
@@ -2,6 +2,8 @@ package service;
 
 import annotation.Auth;
 import domain.Board;
+import exception.NotFoundException;
+import exception.UnauthorizedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import repository.BoardMapper;
@@ -24,7 +26,8 @@ public class BoardServiceImpl implements BoardService{
 
 
     @Override
-    public Board createBoard(Board board) {
+    public Board createBoard(Board board, String token) {
+        board.setWriter_Id(jwtUtil.getUserIdFromJWT(token));
         boardMapper.insertBoard(board);
         return boardMapper.getBoardById(board.getBoard_Id());
     }
@@ -35,23 +38,33 @@ public class BoardServiceImpl implements BoardService{
     }
 
     @Override
-    public Board getBoard(Long boardId) {
+    public Board getBoard(Long boardId) throws Exception{
+        Board select = boardMapper.getBoardById(boardId);
+
+        if(select == null) throw new NotFoundException("존재하지 않는 게시판입니다");
+
         return boardMapper.getBoardById(boardId);
     }
 
-    @Auth
-    @Override
-    public void updateBoard(Board board, Long boardId, String token) {
-        Board selectBoard = boardMapper.getBoardById(boardId);
-        if(jwtUtil.getUserIdFromJWT(token) != selectBoard.getBoard_Id()){
 
+    @Override
+    public void updateBoard(Board board, Long boardId, String token) throws Exception{
+        Board selectBoard = boardMapper.getBoardById(boardId);
+
+        if(selectBoard == null){
+            throw new NotFoundException("존재하지 않는 게시판입니다");
         }
+
+        if(jwtUtil.getUserIdFromJWT(token) != selectBoard.getBoard_Id()){
+            throw new UnauthorizedException("권한이 없습니다");
+        }
+
         board.setBoard_Id(boardId);
         boardMapper.updateBoard(board);
     }
 
     @Override
-    public Board searchBoard(Board board) {
+    public List<Board> searchBoard(Board board) {
         if(board.getDestination() == null){
             return boardMapper.getBoardListByStartingPoint(board);
         }
@@ -65,10 +78,15 @@ public class BoardServiceImpl implements BoardService{
 
 
     @Override
-    public void deleteBoard(Long boardId, String token) {
+    public void deleteBoard(Long boardId, String token) throws Exception{
         Board selectBoard = boardMapper.getBoardById(boardId);
-        if(jwtUtil.getUserIdFromJWT(token) != selectBoard.getBoard_Id()){
 
+        if(selectBoard == null){
+            throw new NotFoundException("존재하지 않는 게시물입니다");
+        }
+
+        if(jwtUtil.getUserIdFromJWT(token) != selectBoard.getBoard_Id()){
+            throw new UnauthorizedException("권한이 없습니다");
         }
         boardMapper.deleteBoard(boardId);
     }
